ui/requirements.txt
gradio>=4.36
requests>=2.31.0
python-dotenv>=1.0.1
pyyaml>=6.0.1